Fix linters issues
