+++ app/views/devise/registrations/new.html.erb
@@ -1,4 +1,4 @@
-<section class="pt-16 w-full h-screen bg-zinc-100 flex flex-col justify-between" id="log-in">
+<section class="pt-16 w-full h-screen bg-zinc-100 flex flex-col justify-between" id="sign-up">
   <div class="container mx-auto h-full flex flex-col">
     <div class="content-row w-full">
       <%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
